requirements.txt
llama-cpp-python==0.2.71
python-dotenv==1.0.1
python-telegram-bot==21.1.1